requirements.txt
Django
pillow
psycopg
dj-database-url